+++ src/app/constants/enchant_item/_enchant_table.ts
@@ -1927,8 +1927,8 @@ export const EnchantTable: EntTable[] = [
     enchants: [
       null,
       ['Spirit_Of_Star_03_CN', 'Spirit_Of_Star_04_CN', 'Spirit_Of_Star_05_CN', 'Spirit_Of_Star_06_CN', 'Spirit_Of_Star_07_CN'],
-      ['Gem_Str_TH_1', 'Gem_Vit_TH_1', 'Gem_Agi_TH_1', 'Gem_Int_TH_1', 'Gem_Dex_TH_1', 'Gem_Luk_TH_1'],
-      ['Gem_Str_TH_1', 'Gem_Vit_TH_1', 'Gem_Agi_TH_1', 'Gem_Int_TH_1', 'Gem_Dex_TH_1', 'Gem_Luk_TH_1'],
+      ['Gem_Str_TH_1', 'Gem_Vit_TH_1', 'Gem_Agi_TH_1', 'Gem_Int_TH_1', 'Gem_Dex_TH_1', 'Gem_Luk_TH_1', 'Gem_Str_TH_2', 'Gem_Vit_TH_2', 'Gem_Agi_TH_2', 'Gem_Int_TH_2', 'Gem_Dex_TH_2', 'Gem_Luk_TH_2'],
+      ['Gem_Str_TH_1', 'Gem_Vit_TH_1', 'Gem_Agi_TH_1', 'Gem_Int_TH_1', 'Gem_Dex_TH_1', 'Gem_Luk_TH_1', 'Gem_Str_TH_2', 'Gem_Vit_TH_2', 'Gem_Agi_TH_2', 'Gem_Int_TH_2', 'Gem_Dex_TH_2', 'Gem_Luk_TH_2'],
     ],
   },
 ];